Zero-defect decode against BDS 1.26.40.8: item system, MoveEntityDelta, PlayerList
CraftingData: recipes travel as eight typed vectors in fixed order plus the existing mix vectors and clean flag; shaped inputs gain a count prefix; the unlocking requirement moves behind presence bools with a zigzag context. Ingredient descriptors are name-addressed now: numeric tag 0=empty/1=named, then a serialize-name string ('name', 'molang', 'item_tag', 'complex_alias') and a name-based body; the numeric int-id descriptor is gone.

CreativeContent: group category is one byte (was le32). Item stacks: the id-0 short-circuit is gone, an empty stack carries all fields zeroed (this was breaking both item packets on vanilla's air group icons).

MoveEntityDelta: the 16-bit flags header becomes per-field presence optionals; generated packet, with the hand partial bridging the currentPosition/SetFlags convenience API onto the generated struct. PlayerList: per-entry tagged variants (tag 1=add/0=remove plus the action byte with inverted numbering); the packet-level action and the trailing trusted-bool loop are gone. Skins: list counts are varints, armSize is a byte, skin color is raw ARGB le32, and the tail gains trusted-as-string plus the new profile hash (Skin.ProfileHash).

MiNET.Client against live BDS 1.26.40.8 now decodes every server-to-client packet with zero errors, zero unknown ids and zero leftover bytes.

+++ src/MiNET/MiNET/Crafting/Recipes.cs
@@ -59,7 +59,7 @@ namespace MiNET.Crafting
 	/// </summary>
 	public class UnlockingRequirement
 	{
-		public byte Context { get; set; } = 1;
+		public int Context { get; set; } = 1;
 		public List<Item> Ingredients { get; set; }
 	}
